requirements.txt
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
